requirements.txt
Adafruit-Blinka==7.2.2
adafruit-circuitpython-busdevice==5.1.8
adafruit-circuitpython-pn532==2.3.7
adafruit-circuitpython-typing==1.4.0
Adafruit-PlatformDetect==3.22.1
Adafruit-PureIO==1.1.9
aioflask==0.4.0
asgiref==3.5.0
asyncio==3.4.3
autopep8==1.6.0
bash==0.6
board==1.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorzero==2.0
cryptography==36.0.2
delegator.py==0.1.1
firebase-admin==5.2.0
Flask==2.1.1
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
goto-statement==1.2
gpiozero==1.6.2
greenlet==1.1.2
greenletio==0.9.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msgpack==1.0.3
pexpect==4.8.0
pn532pi==1.3
proto-plus==1.20.3
protobuf==3.20.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyftdi==0.54.0
pyparsing==3.0.8
pyserial==3.5
pyusb==1.2.1
requests==2.27.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
rsa==4.8
six==1.16.0
spidev==3.5
sysv-ipc==1.1.0
toml==0.10.2
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.1
zipp==3.8.0
